requirements.txt
aniso8601==9.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
firebase-admin==6.2.0
Flask==2.2.2
Flask-RESTful==0.3.9
gcloud==0.18.3
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.1
msgpack==1.0.5
oauth2client==4.1.3
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyparsing==3.1.0
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2022.5
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.2.2
